feat: Update user repository to handle isAdmin flag in Update function
This commit updates the `userRepository` in the `user_repository.go` file to properly handle the `isAdmin` flag in the `Update` function. Previously, the code only set the `isAdmin` flag to `true` if the `updateData` had a truthy value for `IsAdmin`. However, this logic was incorrect as it did not handle cases where the `updateData` had a falsy value for `IsAdmin`. This commit fixes the issue by setting the `isAdmin` flag based on the value of `updateData.IsAdmin` in every request of the `Update` function.